feat(window): add restriction for Anco windows not supporting cross-display drag
This commit adds a restriction that Anco windows do not support cross-display
drag for both resizing and moving.

It also fixes an issue where non-cross-display-draggable windows used an
incorrect rect when the mouse moved to another display during drag. When
the drag end event was triggered (mouse button released), the window rect
was refreshed using a rect relative to the end display, causing the window
position to be incorrect.

Signed-off-by: liuanguang <liuanguang@h-partners.com>